Update dataframe cleaning and exploratory data analysis process
The commit introduces the addition of matplotlib library for data visualization and includes calculation of missing values as a part of dataframe cleanup in  mid-module-assessment file. Null values in the dataframe are also calculated to improve data understanding. Additionally, specific columns with null values have been dropped to compose a cleaner dataframe.